+++ jvm/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Restore support for JRE 11. (fixes [#528](https://github.com/diffplug/selfie/issues/528))
 - snapshots created by `junit.jupiter.api.TestFactory` are no longer garbage-collected (#534)
+- support parallel testing under `junit.jupiter.execution.parallel.enabled=true` (#534)
 
 ## [2.5.2] - 2025-04-28
 ### Fixed